Add plotLDOS
Multiple energyKeys can be passed to plotLDOS.

The second and third keys are plotted in red and blue respectively. All the
other keys, including the first one, are plotted in black. This can be used
to plot the LDOS of spin defects, where one should supply three energyKeys,
corresponding to the bulk states, the spin up defect states, andn the spin
down defect states, respectively.